Remove Google Drive content concurrency limit
