+++ connectors/src/connectors/google_drive/temporal/file/content_memory_telemetry.test.ts
@@ -1,19 +1,11 @@
-import { heartbeat } from "@connectors/lib/temporal";
 import { afterEach, describe, expect, it, vi } from "vitest";
 
-import {
-  GOOGLE_DRIVE_CONTENT_CONCURRENCY,
-  runWithGoogleDriveContentConcurrency,
-} from "./content_concurrency";
+import { runWithGoogleDriveContentMemoryTelemetry } from "./content_memory_telemetry";
 
 const { statsDClient } = vi.hoisted(() => ({
   statsDClient: {
     distribution: vi.fn(),
   },
-}));
-
-vi.mock("@connectors/lib/temporal", () => ({
-  heartbeat: vi.fn().mockResolvedValue(undefined),
 }));
 
 vi.mock("@connectors/types/shared/statsd", () => ({
@@ -27,7 +19,6 @@ const logger = {
 afterEach(() => {
   vi.useRealTimers();
   vi.restoreAllMocks();
-  vi.mocked(heartbeat).mockReset().mockResolvedValue(undefined);
   statsDClient.distribution.mockReset();
   logger.info.mockReset();
 });
@@ -42,7 +33,7 @@ function makeDeferred() {
 }
 
 function runTask<T>(task: () => Promise<T>) {
-  return runWithGoogleDriveContentConcurrency({
+  return runWithGoogleDriveContentMemoryTelemetry({
     task,
     logger,
     fileSizeBytes: 1_024,
@@ -50,83 +41,25 @@ function runTask<T>(task: () => Promise<T>) {
   });
 }
 
-describe("runWithGoogleDriveContentConcurrency", () => {
-  it("runs at most two content tasks concurrently", async () => {
-    const deferredTasks = Array.from(
-      { length: GOOGLE_DRIVE_CONTENT_CONCURRENCY + 1 },
-      makeDeferred
-    );
+describe("runWithGoogleDriveContentMemoryTelemetry", () => {
+  it("does not limit concurrent content tasks", async () => {
+    const deferredTasks = Array.from({ length: 3 }, makeDeferred);
     let activeTasks = 0;
-    let maxActiveTasks = 0;
 
     const tasks = deferredTasks.map((deferredTask) =>
       runTask(async () => {
         activeTasks++;
-        maxActiveTasks = Math.max(maxActiveTasks, activeTasks);
         await deferredTask.promise;
         activeTasks--;
       })
     );
 
-    await expect.poll(() => activeTasks).toBe(GOOGLE_DRIVE_CONTENT_CONCURRENCY);
-    expect(maxActiveTasks).toBe(GOOGLE_DRIVE_CONTENT_CONCURRENCY);
-
-    const firstDeferredTask = deferredTasks[0];
-    if (!firstDeferredTask) {
-      throw new Error("Expected at least one deferred task");
-    }
-    firstDeferredTask.resolve();
-    await expect
-      .poll(() => maxActiveTasks)
-      .toBe(GOOGLE_DRIVE_CONTENT_CONCURRENCY);
+    await expect.poll(() => activeTasks).toBe(deferredTasks.length);
 
     for (const deferredTask of deferredTasks) {
       deferredTask.resolve();
     }
     await Promise.all(tasks);
-
-    expect(maxActiveTasks).toBe(GOOGLE_DRIVE_CONTENT_CONCURRENCY);
-  });
-
-  it("keeps a slot until a running task settles after heartbeat failure", async () => {
-    vi.useFakeTimers();
-    const runningTasks = Array.from(
-      { length: GOOGLE_DRIVE_CONTENT_CONCURRENCY },
-      makeDeferred
-    );
-    const heartbeatError = new Error("Activity cancelled");
-    vi.mocked(heartbeat).mockRejectedValue(heartbeatError);
-    let startedTasks = 0;
-    let replacementStarted = false;
-
-    const tasks = runningTasks.map((deferredTask) =>
-      runTask(async () => {
-        startedTasks++;
-        await deferredTask.promise;
-      })
-    );
-    const taskAssertions = tasks.map((task) =>
-      expect(task).rejects.toBe(heartbeatError)
-    );
-    await vi.waitFor(() => {
-      expect(startedTasks).toBe(GOOGLE_DRIVE_CONTENT_CONCURRENCY);
-    });
-
-    await vi.advanceTimersByTimeAsync(60_000);
-    const replacementTask = runTask(async () => {
-      replacementStarted = true;
-    });
-    await Promise.resolve();
-
-    expect(replacementStarted).toBe(false);
-
-    for (const runningTask of runningTasks) {
-      runningTask.resolve();
-    }
-    await Promise.all(taskAssertions);
-    await replacementTask;
-
-    expect(replacementStarted).toBe(true);
   });
 
   it("records file metadata and process memory while content is processed", async () => {
@@ -158,7 +91,7 @@ describe("runWithGoogleDriveContentConcurrency", () => {
       .mockReturnValueOnce(sampledMemory)
       .mockReturnValueOnce(memoryAtEnd);
 
-    const taskPromise = runWithGoogleDriveContentConcurrency({
+    const taskPromise = runWithGoogleDriveContentMemoryTelemetry({
       task: () => task.promise,
       logger,
       fileSizeBytes: 42_000,
@@ -172,7 +105,6 @@ describe("runWithGoogleDriveContentConcurrency", () => {
       {
         fileSizeBytes: 42_000,
         mimeType: "application/pdf",
-        queueWaitMs: 0,
         processingDurationMs: 1_000,
         outcome: "success",
         activeOperationsAtStart: 1,
